+++ src/libsuc/CacheCore.h
@@ -34,27 +34,27 @@ Temple Place - Suite 330, Boston, MA 02111-1307, USA.
 enum    ReplacementPolicy  {LRU, RANDOM};
 
 #ifdef SESC_ENERGY
-template<class State, class Addr_t = uint, bool Energy=true>
+template<class State, class Addr_t = uint32_t, bool Energy=true>
 #else
-  template<class State, class Addr_t = uint, bool Energy=false>
+  template<class State, class Addr_t = uint32_t, bool Energy=false>
 #endif
   class CacheGeneric {
   private:
-  static const int STR_BUF_SIZE=1024;
+  static const int32_t STR_BUF_SIZE=1024;
  
   static PowerGroup getRightStat(const char* type);
 
   protected:
-  const uint  size;
-  const uint  lineSize;
-  const uint  addrUnit; //Addressable unit: for most caches = 1 byte
-  const uint  assoc;
-  const uint  log2Assoc;
-  const uint  log2AddrLs;
-  const uint  maskAssoc;
-  const uint  sets;
-  const uint  maskSets;
-  const uint  numLines;
+  const uint32_t  size;
+  const uint32_t  lineSize;
+  const uint32_t  addrUnit; //Addressable unit: for most caches = 1 byte
+  const uint32_t  assoc;
+  const uint32_t  log2Assoc;
+  const uint64_t  log2AddrLs;
+  const uint64_t  maskAssoc;
+  const uint32_t  sets;
+  const uint32_t  maskSets;
+  const uint32_t  numLines;
 
   GStatsEnergy *rdEnergy[2]; // 0 hit, 1 miss
   GStatsEnergy *wrEnergy[2]; // 0 hit, 1 miss
@@ -83,7 +83,7 @@ template<class State, class Addr_t = uint, bool Energy=true>
   virtual CacheLine *findLinePrivate(Addr_t addr)=0;
   protected:
 
-  CacheGeneric(uint s, uint a, uint b, uint u)
+  CacheGeneric(uint32_t s, uint32_t a, uint32_t b, uint32_t u)
   : size(s)
   ,lineSize(b)
   ,addrUnit(u)
@@ -105,7 +105,7 @@ template<class State, class Addr_t = uint, bool Energy=true>
 
   public:
   // Do not use this interface, use other create
-  static CacheGeneric<State, Addr_t, Energy> *create(int size, int assoc, int blksize, int addrUnit, const char *pStr, bool skew);
+  static CacheGeneric<State, Addr_t, Energy> *create(int32_t size, int32_t assoc, int32_t blksize, int32_t addrUnit, const char *pStr, bool skew);
   static CacheGeneric<State, Addr_t, Energy> *create(const char *section, const char *append, const char *format, ...);
   void destroy() {
     delete this;
@@ -122,7 +122,7 @@ template<class State, class Addr_t = uint, bool Energy=true>
   // cleaner interface so that Cache.cpp does not touch the internals.
   //
   // Access the line directly without checking TAG
-  virtual CacheLine *getPLine(uint l) = 0;
+  virtual CacheLine *getPLine(uint32_t l) = 0;
 
   //ALL USERS OF THIS CLASS PLEASE READ:
   //
@@ -209,30 +209,30 @@ template<class State, class Addr_t = uint, bool Energy=true>
     return l;
   }
 
-  uint  getLineSize() const   { return lineSize;    }
-  uint  getAssoc() const      { return assoc;       }
-  uint  getLog2AddrLs() const { return log2AddrLs;  }
-  uint  getLog2Assoc() const  { return log2Assoc;   }
-  uint  getMaskSets() const   { return maskSets;    }
-  uint  getNumLines() const   { return numLines;    }
-  uint  getNumSets() const    { return sets;        }
+  uint32_t  getLineSize() const   { return lineSize;    }
+  uint32_t  getAssoc() const      { return assoc;       }
+  uint32_t  getLog2AddrLs() const { return log2AddrLs;  }
+  uint32_t  getLog2Assoc() const  { return log2Assoc;   }
+  uint32_t  getMaskSets() const   { return maskSets;    }
+  uint32_t  getNumLines() const   { return numLines;    }
+  uint32_t  getNumSets() const    { return sets;        }
 
   Addr_t calcTag(Addr_t addr)       const { return (addr >> log2AddrLs);              }
 
-  uint calcSet4Tag(Addr_t tag)     const { return (tag & maskSets);                  }
-  uint calcSet4Addr(Addr_t addr)   const { return calcSet4Tag(calcTag(addr));        }
+  uint32_t calcSet4Tag(Addr_t tag)     const { return (tag & maskSets);                  }
+  uint32_t calcSet4Addr(Addr_t addr)   const { return calcSet4Tag(calcTag(addr));        }
 
-  uint calcIndex4Set(uint set)    const { return (set << log2Assoc);                }
-  uint calcIndex4Tag(uint tag)    const { return calcIndex4Set(calcSet4Tag(tag));   }
-  uint calcIndex4Addr(Addr_t addr) const { return calcIndex4Set(calcSet4Addr(addr)); }
+  uint32_t calcIndex4Set(uint32_t set)    const { return (set << log2Assoc);                }
+  uint32_t calcIndex4Tag(uint32_t tag)    const { return calcIndex4Set(calcSet4Tag(tag));   }
+  uint32_t calcIndex4Addr(Addr_t addr) const { return calcIndex4Set(calcSet4Addr(addr)); }
 
   Addr_t calcAddr4Tag(Addr_t tag)   const { return (tag << log2AddrLs);                   }
 };
 
 #ifdef SESC_ENERGY
-template<class State, class Addr_t = uint, bool Energy=true>
+template<class State, class Addr_t = uint32_t, bool Energy=true>
 #else
-template<class State, class Addr_t = uint, bool Energy=false>
+template<class State, class Addr_t = uint32_t, bool Energy=false>
 #endif
 class CacheAssoc : public CacheGeneric<State, Addr_t, Energy> {
   using CacheGeneric<State, Addr_t, Energy>::numLines;
@@ -252,7 +252,7 @@ protected:
   ReplacementPolicy policy;
 
   friend class CacheGeneric<State, Addr_t, Energy>;
-  CacheAssoc(int size, int assoc, int blksize, int addrUnit, const char *pStr);
+  CacheAssoc(int32_t size, int32_t assoc, int32_t blksize, int32_t addrUnit, const char *pStr);
 
   Line *findLinePrivate(Addr_t addr);
 public:
@@ -262,7 +262,7 @@ public:
   }
 
   // TODO: do an iterator. not this junk!!
-  Line *getPLine(uint l) {
+  Line *getPLine(uint32_t l) {
     // Lines [l..l+assoc] belong to the same set
     I(l<numLines);
     return content[l];
@@ -272,9 +272,9 @@ public:
 };
 
 #ifdef SESC_ENERGY
-template<class State, class Addr_t = uint, bool Energy=true>
+template<class State, class Addr_t = uint32_t, bool Energy=true>
 #else
-template<class State, class Addr_t = uint, bool Energy=false>
+template<class State, class Addr_t = uint32_t, bool Energy=false>
 #endif
 class CacheDM : public CacheGeneric<State, Addr_t, Energy> {
   using CacheGeneric<State, Addr_t, Energy>::numLines;
@@ -290,7 +290,7 @@ protected:
   Line **content;
 
   friend class CacheGeneric<State, Addr_t, Energy>;
-  CacheDM(int size, int blksize, int addrUnit, const char *pStr);
+  CacheDM(int32_t size, int32_t blksize, int32_t addrUnit, const char *pStr);
 
   Line *findLinePrivate(Addr_t addr);
 public:
@@ -300,7 +300,7 @@ public:
   };
 
   // TODO: do an iterator. not this junk!!
-  Line *getPLine(uint l) {
+  Line *getPLine(uint32_t l) {
     // Lines [l..l+assoc] belong to the same set
     I(l<numLines);
     return content[l];
@@ -310,9 +310,9 @@ public:
 };
 
 #ifdef SESC_ENERGY
-template<class State, class Addr_t = uint, bool Energy=true>
+template<class State, class Addr_t = uint32_t, bool Energy=true>
 #else
-template<class State, class Addr_t = uint, bool Energy=false>
+template<class State, class Addr_t = uint32_t, bool Energy=false>
 #endif
 class CacheDMSkew : public CacheGeneric<State, Addr_t, Energy> {
   using CacheGeneric<State, Addr_t, Energy>::numLines;
@@ -328,7 +328,7 @@ protected:
   Line **content;
 
   friend class CacheGeneric<State, Addr_t, Energy>;
-  CacheDMSkew(int size, int blksize, int addrUnit, const char *pStr);
+  CacheDMSkew(int32_t size, int32_t blksize, int32_t addrUnit, const char *pStr);
 
   Line *findLinePrivate(Addr_t addr);
 public:
@@ -338,7 +338,7 @@ public:
   };
 
   // TODO: do an iterator. not this junk!!
-  Line *getPLine(uint l) {
+  Line *getPLine(uint32_t l) {
     // Lines [l..l+assoc] belong to the same set
     I(l<numLines);
     return content[l];
@@ -348,7 +348,7 @@ public:
 };
 
 
-template<class Addr_t=uint>
+template<class Addr_t=uint32_t>
 class StateGeneric {
 private:
   Addr_t tag;
